pyproject.toml
[project]
name = "morning-radio"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "edge-tts>=7.0.0",
    "feedgen>=1.0.0",
    "jinja2>=3.1.5",
    "jupyter>=1.1.1",
    "llm-openrouter>=0.3",
    "llm>=0.19.1",
    "openmeteo-requests>=1.3.0",
    "pandas>=2.2.3",
    "quarto-cli>=1.6.40",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
    "schedule>=1.2.2",
    "yfinance>=0.2.52",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "llama-cpp-python>=0.3.6",
    "llvmlite>=0.44.0",
    "pydub>=0.25.1",
    "audioop-lts>=0.2.1",
    "kokoro-onnx",
    "soundfile>=0.13.0",
]

[tool.uv.sources]
kokoro-onnx = { git = "https://github.com/thewh1teagle/kokoro-onnx" }
